requirements.txt
customtkinter>=5.2.1
openpyxl>=3.1.2
python-dotenv>=1.0.1
